requirements.txt
pyTermTk==0.10.22a0
